Refactor gemini open responses into generic module so can be used by other providers in future
