+++ crates/agentgateway/src/llm/conversion/mod.rs
@@ -2,5 +2,6 @@ pub mod bedrock;
 pub mod completions;
 pub mod gemini;
 pub mod messages;
+pub mod openai_compat;
 pub mod responses;
 pub mod vertex;
