+++ crates/agentgateway/src/llm/types/responses.rs
@@ -382,8 +382,8 @@ impl RequestType for Request {
 		serde_json::to_vec(&self).map_err(AIError::RequestMarshal)
 	}
 
-	fn to_gemini(&self) -> Result<Vec<u8>, AIError> {
-		conversion::gemini::from_responses::translate(self)
+	fn to_openai_chat_completions(&self) -> Result<Vec<u8>, AIError> {
+		conversion::openai_compat::from_responses::translate(self)
 	}
 
 	fn to_bedrock(
